+++ src/main/java/com/yuo/endless/Proxy/IProxy.java
@@ -1,0 +1,5 @@
+package com.yuo.endless.Proxy;
+
+public interface IProxy {
+    default void registerHandlers() {}
+}
